maybe correct payload

+++ src/main/java/tcpserver/Backend/MQTT_PubPayload.java
@@ -59,7 +59,7 @@ public class MQTT_PubPayload {
         this.seq = seq;
     }
 
-    public void setData(float lat, float lon) {
+    public void setData(int lat, int lon) {
         data.setData(lat, lon);
     }
 
@@ -69,21 +69,21 @@ public class MQTT_PubPayload {
 }
 
 class GPS_data {
-    private float lat = 0;
-    private float lon = 0;
+    private int lat = 0;
+    private int lon = 0;
 
     public GPS_data() {}
 
-    public void setData(float lat, float lon) {
+    public void setData(int lat, int lon) {
         this.lat = lat;
         this.lon = lon;
     }
 
-    public float getLat() {
+    public int getLat() {
         return lat;
     }
 
-    public float getLon() {
+    public int getLon() {
         return lon;
     }
 }
